+++ CHANGELOG.md
@@ -6,6 +6,7 @@ CHANGELOG
 * Add files allocation functionality to the transform command.
 * Rename example files (add locale to the file name).
 * Add alias for the Lint YAML command: `aeliot_trans_maintain:yaml:lint`.
+* Move commands which works with YAML files to the separate namespace.
 
 2.5.1
 -----