+++ CHANGELOG.md
@@ -9,8 +9,9 @@ CHANGELOG
 * Move commands which works with YAML files to the separate namespace.
 * Add detector (linter) of empty values of translations.
 * Move public constants of `\Aeliot\Bundle\TransMaintain\Service\Yaml\LinterRegistry` to `\Aeliot\Bundle\TransMaintain\Service\Yaml\Linter\LinterInterface`.
-* Simplify report line definitions through the defining of columns order and headers in the single place of each report line model. 
+* Simplify report line definitions through the defining of columns order and headers in the single place of each report line model.
 * Make Google Translation model configurable.
+* Use `.env` files to load parameters for the development kernel.
 
 2.5.1
 -----